+++ interfaces/otlp/tests/unit/test_endpoints.py
@@ -4,39 +4,42 @@
 """Feature: OTLP endpoint handling."""
 
 import json
-from typing import Any, cast
-from unittest.mock import patch
+from collections.abc import Sequence
+from typing import Any, Final, Literal
 
 import ops
 import pytest
 from ops import testing
 from ops.testing import Relation, State
 
-from charmlibs.interfaces.otlp._otlp import OtlpProviderAppData, _OtlpEndpoint
+from charmlibs.interfaces.otlp._otlp import DEFAULT_PROVIDER_RELATION_NAME as RECEIVE
+from charmlibs.interfaces.otlp._otlp import DEFAULT_REQUIRER_RELATION_NAME as SEND
+from charmlibs.interfaces.otlp._otlp import (
+    OtlpProvider,
+    OtlpRequirer,
+    _OtlpEndpoint,
+    _OtlpProviderAppData,
+)
+from conftest import ALL_PROTOCOLS, ALL_TELEMETRIES
 
-ALL_PROTOCOLS = ['grpc', 'http']
-ALL_TELEMETRIES = ['logs', 'metrics', 'traces']
-EMPTY_REQUIRER = {
-    'rules': json.dumps({'logql': {}, 'promql': {}}),
-    'metadata': json.dumps({}),
-}
-
-RECEIVE_OTLP = Relation('receive-otlp', remote_app_data=EMPTY_REQUIRER)
+PROTOCOLS: Final[list[Literal['http', 'grpc']]] = ['http', 'grpc']
+TELEMETRIES: Final[list[Literal['metrics', 'logs']]] = ['metrics', 'logs']
 
 
-def test_new_endpoint_key_is_ignored_by_databag_model() -> None:
+def test_new_endpoint_key_is_ignored_by_databag_model():
     # GIVEN the provider offers a new endpoint type (protocol or telemetry)
     # * the requirer does not support this new endpoint type
     endpoint = {
         'protocol': 'new_protocol',
         'endpoint': 'http://host:4317',
         'telemetries': ['logs'],
+        'insecure': False,
         'new_key': 'value',
     }
 
     # WHEN validating the provider databag model, which the requirer uses to access endpoints
     # THEN the validation succeeds
-    provider_databag: OtlpProviderAppData = OtlpProviderAppData.model_validate({
+    provider_databag: _OtlpProviderAppData = _OtlpProviderAppData.model_validate({
         'endpoints': [endpoint]
     })
     assert provider_databag
@@ -56,11 +59,13 @@ def test_new_endpoint_key_is_ignored_by_databag_model() -> None:
                         'protocol': 'new_protocol',
                         'endpoint': 'http://host:0000',
                         'telemetries': ['metrics'],
+                        'insecure': False,
                     },
                     {
                         'protocol': 'http',
                         'endpoint': 'http://host:4317',
                         'telemetries': ['metrics'],
+                        'insecure': False,
                     },
                 ]),
             },
@@ -78,13 +83,12 @@ def test_new_endpoint_key_is_ignored_by_databag_model() -> None:
                         'protocol': 'http',
                         'endpoint': 'http://host:4317',
                         'telemetries': ['logs', 'new_telemetry', 'traces'],
+                        'insecure': False,
                     },
                 ]),
             },
             _OtlpEndpoint(
-                protocol='http',
-                endpoint='http://host:4317',
-                telemetries=['logs', 'traces'],
+                protocol='http', endpoint='http://host:4317', telemetries=['logs', 'traces']
             ),
         ),
         (
@@ -96,6 +100,7 @@ def test_new_endpoint_key_is_ignored_by_databag_model() -> None:
                         'protocol': 'http',
                         'endpoint': 'http://host:4317',
                         'telemetries': ['metrics'],
+                        'insecure': False,
                     }
                 ]),
                 'does_not': '"exist"',
@@ -115,22 +120,17 @@ def test_send_otlp_invalid_databag(
 ):
     # GIVEN a remote app provides an _OtlpEndpoint
     # WHEN they are related over the "send-otlp" endpoint
-    provider = Relation('send-otlp', id=123, remote_app_data=provides)
+    provider = Relation(SEND, id=123, remote_app_data=provides)
     state = State(relations=[provider], leader=True)
 
     with otlp_requirer_ctx(otlp_requirer_ctx.on.update_status(), state=state) as mgr:
         # WHEN the requirer processes the relation data
         # * the requirer supports all protocols and telemetries
-        charm_any = cast('Any', mgr.charm)
-        with (
-            patch.object(charm_any.otlp_requirer, '_protocols', new=ALL_PROTOCOLS),
-            patch.object(charm_any.otlp_requirer, '_telemetries', new=ALL_TELEMETRIES),
-        ):
-            # THEN the requirer does not raise an error
-            # * the returned endpoint does not include new protocols or telemetries
-            assert mgr.run()
-            result = charm_any.otlp_requirer.endpoints[123]
-            assert result.model_dump() == otlp_endpoint.model_dump()
+        # THEN the requirer does not raise an error
+        # * the returned endpoint does not include new protocols or telemetries
+        assert mgr.run()
+        endpoints = OtlpRequirer(mgr.charm, SEND, ALL_PROTOCOLS, ALL_TELEMETRIES).endpoints
+        assert endpoints[123].model_dump() == otlp_endpoint.model_dump()
 
 
 @pytest.mark.parametrize(
@@ -142,13 +142,11 @@ def test_send_otlp_invalid_databag(
             {
                 123: _OtlpEndpoint(
                     protocol='http',
-                    endpoint='http://provider-123.endpoint:4318',
+                    endpoint='http://provider-123:4318',
                     telemetries=['logs', 'metrics'],
                 ),
                 456: _OtlpEndpoint(
-                    protocol='grpc',
-                    endpoint='http://provider-456.endpoint:4317',
-                    telemetries=['traces'],
+                    protocol='grpc', endpoint='http://provider-456:4317', telemetries=['traces']
                 ),
             },
         ),
@@ -157,9 +155,7 @@ def test_send_otlp_invalid_databag(
             ALL_TELEMETRIES,
             {
                 456: _OtlpEndpoint(
-                    protocol='grpc',
-                    endpoint='http://provider-456.endpoint:4317',
-                    telemetries=['traces'],
+                    protocol='grpc', endpoint='http://provider-456:4317', telemetries=['traces']
                 )
             },
         ),
@@ -168,14 +164,10 @@ def test_send_otlp_invalid_databag(
             ['metrics'],
             {
                 123: _OtlpEndpoint(
-                    protocol='http',
-                    endpoint='http://provider-123.endpoint:4318',
-                    telemetries=['metrics'],
+                    protocol='http', endpoint='http://provider-123:4318', telemetries=['metrics']
                 ),
                 456: _OtlpEndpoint(
-                    protocol='http',
-                    endpoint='http://provider-456.endpoint:4318',
-                    telemetries=['metrics'],
+                    protocol='http', endpoint='http://provider-456:4318', telemetries=['metrics']
                 ),
             },
         ),
@@ -184,8 +176,8 @@ def test_send_otlp_invalid_databag(
 )
 def test_send_otlp_with_varying_requirer_support(
     otlp_requirer_ctx: testing.Context[ops.CharmBase],
-    protocols: list[str],
-    telemetries: list[str],
+    protocols: Sequence[Literal['http', 'grpc']],
+    telemetries: Sequence[Literal['logs', 'metrics', 'traces']],
     expected: dict[int, _OtlpEndpoint],
 ):
     # GIVEN a remote app provides multiple _OtlpEndpoints
@@ -193,8 +185,9 @@ def test_send_otlp_with_varying_requirer_support(
         'endpoints': json.dumps([
             {
                 'protocol': 'http',
-                'endpoint': 'http://provider-123.endpoint:4318',
+                'endpoint': 'http://provider-123:4318',
                 'telemetries': ['logs', 'metrics'],
+                'insecure': False,
             }
         ])
     }
@@ -202,41 +195,27 @@ def test_send_otlp_with_varying_requirer_support(
         'endpoints': json.dumps([
             {
                 'protocol': 'grpc',
-                'endpoint': 'http://provider-456.endpoint:4317',
+                'endpoint': 'http://provider-456:4317',
                 'telemetries': ['traces'],
+                'insecure': False,
             },
             {
                 'protocol': 'http',
-                'endpoint': 'http://provider-456.endpoint:4318',
+                'endpoint': 'http://provider-456:4318',
                 'telemetries': ['metrics'],
+                'insecure': False,
             },
         ])
     }
 
     # WHEN they are related over the "send-otlp" endpoint
-    provider_0 = Relation(
-        'send-otlp',
-        id=123,
-        remote_app_data=remote_app_data_1,
-    )
-    provider_1 = Relation(
-        'send-otlp',
-        id=456,
-        remote_app_data=remote_app_data_2,
-    )
-    state = State(
-        relations=[provider_0, provider_1],
-        leader=True,
-    )
+    provider_0 = Relation(SEND, id=123, remote_app_data=remote_app_data_1)
+    provider_1 = Relation(SEND, id=456, remote_app_data=remote_app_data_2)
+    state = State(relations=[provider_0, provider_1], leader=True)
 
     # AND WHEN the requirer has varying support for OTLP protocols and telemetries
     with otlp_requirer_ctx(otlp_requirer_ctx.on.update_status(), state=state) as mgr:
-        charm_any = cast('Any', mgr.charm)
-        with (
-            patch.object(charm_any.otlp_requirer, '_protocols', new=protocols),
-            patch.object(charm_any.otlp_requirer, '_telemetries', new=telemetries),
-        ):
-            remote_endpoints = charm_any.otlp_requirer.endpoints
+        remote_endpoints = OtlpRequirer(mgr.charm, SEND, protocols, telemetries).endpoints
 
     # THEN the returned endpoints are filtered accordingly
     assert {k: v.model_dump() for k, v in remote_endpoints.items()} == {
@@ -250,8 +229,9 @@ def test_send_otlp(otlp_requirer_ctx: testing.Context[ops.CharmBase]):
         'endpoints': json.dumps([
             {
                 'protocol': 'http',
-                'endpoint': 'http://provider-123.endpoint:4318',
+                'endpoint': 'http://provider-123:4318',
                 'telemetries': ['logs', 'metrics'],
+                'insecure': False,
             }
         ])
     }
@@ -259,50 +239,36 @@ def test_send_otlp(otlp_requirer_ctx: testing.Context[ops.CharmBase]):
         'endpoints': json.dumps([
             {
                 'protocol': 'grpc',
-                'endpoint': 'http://provider-456.endpoint:4317',
+                'endpoint': 'http://provider-456:4317',
                 'telemetries': ['traces'],
+                'insecure': False,
             },
             {
                 'protocol': 'http',
-                'endpoint': 'http://provider-456.endpoint:4318',
+                'endpoint': 'http://provider-456:4318',
                 'telemetries': ['metrics'],
+                'insecure': False,
             },
         ])
     }
 
     expected_endpoints = {
         456: _OtlpEndpoint(
-            protocol='http',
-            endpoint='http://provider-456.endpoint:4318',
-            telemetries=['metrics'],
+            protocol='http', endpoint='http://provider-456:4318', telemetries=['metrics']
         ),
         123: _OtlpEndpoint(
-            protocol='http',
-            endpoint='http://provider-123.endpoint:4318',
-            telemetries=['logs', 'metrics'],
+            protocol='http', endpoint='http://provider-123:4318', telemetries=['logs', 'metrics']
         ),
     }
 
     # WHEN they are related over the "send-otlp" endpoint
-    provider_1 = Relation(
-        'send-otlp',
-        id=123,
-        remote_app_data=remote_app_data_1,
-    )
-    provider_2 = Relation(
-        'send-otlp',
-        id=456,
-        remote_app_data=remote_app_data_2,
-    )
-    state = State(
-        relations=[provider_1, provider_2],
-        leader=True,
-    )
+    provider_1 = Relation(SEND, id=123, remote_app_data=remote_app_data_1)
+    provider_2 = Relation(SEND, id=456, remote_app_data=remote_app_data_2)
+    state = State(relations=[provider_1, provider_2], leader=True)
 
     # AND WHEN otelcol supports a subset of OTLP protocols and telemetries
     with otlp_requirer_ctx(otlp_requirer_ctx.on.update_status(), state=state) as mgr:
-        charm_any = cast('Any', mgr.charm)
-        remote_endpoints = charm_any.otlp_requirer.endpoints
+        remote_endpoints = OtlpRequirer(mgr.charm, SEND, PROTOCOLS, TELEMETRIES).endpoints
 
     # THEN the returned endpoints are filtered accordingly
     assert {k: v.model_dump() for k, v in remote_endpoints.items()} == {
@@ -312,10 +278,14 @@ def test_send_otlp(otlp_requirer_ctx: testing.Context[ops.CharmBase]):
 
 def test_receive_otlp(otlp_provider_ctx: testing.Context[ops.CharmBase]):
     # GIVEN a receive-otlp relation
-    state = State(
-        leader=True,
-        relations=[RECEIVE_OTLP],
+    receiver = Relation(
+        RECEIVE,
+        remote_app_data={
+            'rules': json.dumps({'logql': {}, 'promql': {}}),
+            'metadata': '{}',
+        },
     )
+    state = State(leader=True, relations=[receiver])
 
     # AND WHEN any event executes the reconciler
     state_out = otlp_provider_ctx.run(otlp_provider_ctx.on.update_status(), state=state)
@@ -328,12 +298,13 @@ def test_receive_otlp(otlp_provider_ctx: testing.Context[ops.CharmBase]):
                 'protocol': 'http',
                 'endpoint': 'http://fqdn:4318',
                 'telemetries': ['metrics'],
+                'insecure': False,
             }
         ],
     }
     assert (actual_endpoints := json.loads(local_app_data.get('endpoints', '[]')))
     assert (
-        OtlpProviderAppData.model_validate({'endpoints': actual_endpoints}).model_dump()
+        _OtlpProviderAppData.model_validate({'endpoints': actual_endpoints}).model_dump()
         == expected_endpoints
     )
 
@@ -393,10 +364,38 @@ def test_favor_modern_endpoints(
     # GIVEN a list of endpoints
     state = State(leader=True)
     with otlp_requirer_ctx(otlp_requirer_ctx.on.update_status(), state=state) as mgr:
-        charm_any = cast('Any', mgr.charm)
-
         # WHEN the requirer selects an endpoint
-        result = charm_any.otlp_requirer._favor_modern_endpoints(endpoints)
+        requirer = OtlpRequirer(mgr.charm, SEND, PROTOCOLS, TELEMETRIES)
 
     # THEN the most modern one is chosen
-    assert result.protocol == expected_protocol
+    assert requirer._favor_modern_endpoints(endpoints).protocol == expected_protocol
+
+
+def test_add_endpoint_insecure(otlp_requirer_ctx: testing.Context[ops.CharmBase]):
+    state = State(leader=True)
+    with otlp_requirer_ctx(otlp_requirer_ctx.on.update_status(), state=state) as mgr:
+        # GIVEN a provider
+        provider = OtlpProvider(mgr.charm)
+
+    # WHEN insecure and secure endpoints are added
+    provider.add_endpoint(
+        protocol='http', endpoint='http://host:4318', telemetries=['metrics'], insecure=True
+    ).add_endpoint(
+        protocol='grpc', endpoint='http://host:4317', telemetries=['traces'], insecure=False
+    )
+
+    # THEN their security settings are preserved
+    assert provider._endpoints == [
+        _OtlpEndpoint(
+            protocol='http',
+            endpoint='http://host:4318',
+            telemetries=['metrics'],
+            insecure=True,
+        ),
+        _OtlpEndpoint(
+            protocol='grpc',
+            endpoint='http://host:4317',
+            telemetries=['traces'],
+            insecure=False,
+        ),
+    ]